feat!: Update Address transactions to fix SonarQube issue
[#OCD-3587]
